GH-16
Add leagues page.

Tbh I dont think this is ready for a pull request because there is not much functionality actually here lol. Mostly waiting for the server to be ready to handle leagues... if it is, I couldn't find the code anywhere so feel free to lmk where it is :).